Restore git global config
`setup-git-redirect.sh` is needed for tests, because go needs SSH
authentication, when downloading modules dependencies. The redirect
script, though, does not revert the global changes done to the git
config (in that case: SSH instead of HTTPS). The `checkout` action,
on the other hand, needs the HTTPS for checkout, because the SSH key at the
beginning of the action is no longer available, it was properly cleaned
up in the `webfactory/ssh-agent` post clean up.

Implementing workaround as in https://github.com/webfactory/ssh-agent/issues/184